Add padding to main page header
There's no padding applied to the main page header box, so the
descenders (as in 'p' and 'y') appear to be touching the box border,
which is undesirable (this doesn't happen at the top when the default
font, Verdana, is applied because it leaves extra space above characters
for ascenders, such as accents). Apply padding to fix this.