Fix credits/contributor rendering issues
Some larger code refactoring is probably in order, but a few fixes until
somebody gets around to doing that:

- On sprite pages, properly display the username when no contact info
is provided (instead of displaying "undefined").
- On sprite pages and the contributors page, properly escape HTML, such
as in anchor hrefs to Discord userid tags.
- On the contributors page, properly handle users with no username
provided (an italicized "Anonymous"), instead of hard-coding a list of
IDs to exclude from the list.
- Unify contact info rendering so it works similary on both the sprite
pages and the contributors page.
- Handle various other edge cases more exhaustively.